+++ src/modules/components/loading-spinner/index.tsx
@@ -1,9 +1,13 @@
 import { Root, CircularProgressIcon } from "./style";
 
-const LoadingSpinner = () => {
+interface ILoadingSpinnerProps {
+  color?: string;
+}
+
+const LoadingSpinner = ({ color }: ILoadingSpinnerProps) => {
   return (
     <Root container justifyContent="center" alignItems="center">
-      <CircularProgressIcon size={60} />
+      <CircularProgressIcon color={color} size={60} />
     </Root>
   );
 };